refactor(image): 파일 경로 단순 수정
