+++ immilog-core/src/main/java/com/backend/immilog/image/application/usecase/UploadImageUseCase.kt
@@ -6,7 +6,7 @@ import com.backend.immilog.image.domain.enums.ImageType
 import com.backend.immilog.image.domain.model.Image
 import com.backend.immilog.image.domain.model.ImageMetadata
 import com.backend.immilog.image.domain.model.ImagePath
-import com.backend.immilog.image.infrastructure.gateway.FileStorageHandler
+import com.backend.immilog.image.domain.service.FileStorageHandler
 import com.backend.immilog.shared.application.event.DomainEventPublisher
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
